Rename component to Freeze
