+++ example/src/Examples/API/Freeze.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import {
   Canvas,
   Group,
@@ -6,25 +6,32 @@ import {
   Text,
   useClockValue,
   useDerivedValue,
-  ImageDrawing,
+  Freeze,
   rect,
 } from "@shopify/react-native-skia";
 
 const size = 200;
-const n = 91;
+const n = 3;
 
 export const FreezeExample = () => {
+  const [color, setColor] = useState("black");
   const clock = useClockValue();
   const transform = useDerivedValue(
     () => [{ translateY: 100 }, { rotate: (Math.PI * clock.current) / 4000 }],
     [clock]
   );
+  useEffect(() => {
+    const h = setInterval(() => {
+      setColor("#" + (Math.random().toString(16) + "00000").slice(2, 8));
+    }, 10000);
+    return () => clearInterval(h);
+  }, []);
   return (
     <Canvas style={{ flex: 1, margin: 50 }} debug>
       <Group origin={{ x: size / 2, y: size / 2 }} transform={transform}>
-        <ImageDrawing rect={rect(0, 0, 200, 200)}>
-          <Checkerboard />
-        </ImageDrawing>
+        <Freeze rect={rect(0, 0, 200, 200)}>
+          <Checkerboard color={color} />
+        </Freeze>
       </Group>
       <Text
         x={20}
@@ -37,7 +44,7 @@ export const FreezeExample = () => {
   );
 };
 
-const Checkerboard = () => {
+const Checkerboard = ({ color }: { color: string }) => {
   // draw a n * n checkerboard
   return (
     <>
@@ -48,7 +55,7 @@ const Checkerboard = () => {
           y={(Math.floor(i / n) * size) / n}
           width={size / n}
           height={size / n}
-          color={i % 2 ? "#000" : "#ddd"}
+          color={i % 2 ? color : "#ddd"}
         />
       ))}
     </>
