+++ package/src/renderer/components/image/index.ts
@@ -1,5 +1,4 @@
 export * from "./Image";
 export * from "./ImageShader";
 export * from "./ImageSVG";
-export * from "./ImageDrawing";
 export * from "./BoxFit";
